CR comments fixes
