+++ modules/components/documents/component.php
@@ -67,10 +67,11 @@ class Component extends Document {
 		return new Component_Overridable_Props( $meta );
 	}
 
-	public function set_overridable_props( $data ): Parse_Result {
+	public function update_overridable_props( $data ): Parse_Result {
 		$parser = Component_Overridable_Props_Parser::make();
 
 		$result = $parser->parse( $data );
+		$result = $parser->parse( [] );
 
 		if ( ! $result->is_valid() ) {
 			return $result;
